Fix syntax highlighting for #o and #b literals
Octal and binary literals were being highlighted as errors unless
followed by a newline (so for example, when followed by a space or a
')'). This commit introduces the intended behavior by adding the "end of
word" marker to the end of the appropriate regexes.

My testing indicates that this change yields correct behavior, but I do
not fully grok the surrounding regexes, so no guarantees I'm afraid.